rpi/requirements.txt
Flask==1.1.2
PyDispatcher~=2.0.5